final version

+++ tp2/comando.h
@@ -7,7 +7,7 @@ typedef struct comando {
 	bool es_valido;
 	char* nombre;
 	char** parametros;
-	int cantidad_parametros;
+	size_t cantidad_parametros;
 } comando_t;
 
 #endif
